requirements.txt
requests
proxmoxer
python-dotenv
